+++ Changelog.md
@@ -1,6 +1,7 @@
 # Next
 
 - Changes `Moya.Error.Underlying` to have `NSError` instead of `ErrorType`. *Breaking change*
+- Implements inflights tracking by adding `trackInflights = true` to your provider. *Breaking change*
 
 # 6.4.0