+++ Changelog.md
@@ -1,8 +1,8 @@
 # Next
 
-- Changes `Moya.Error.Underlying` to have `NSError` instead of `ErrorType`. *Breaking change*
-- Implements inflights tracking by adding `trackInflights = true` to your provider. *Breaking change*
-- Changes `MoyaProvider.RequestClosure` to have `Result<NSURLRequest, Moya.Error> -> Void` instead of `NSURLRequest -> Void` as a `done` closure parameter. *Breaking change*
+- **Breaking Change** Changes `Moya.Error.Underlying` to have `NSError` instead of `ErrorType`.
+- **Breaking Change** Implements inflights tracking by adding `trackInflights = true` to your provider.
+- **Breaking Change** Changes `MoyaProvider.RequestClosure` to have `Result<NSURLRequest, Moya.Error> -> Void` instead of `NSURLRequest -> Void` as a `done` closure parameter.
 
 # 6.4.0