+++ Changelog.md
@@ -1,4 +1,6 @@
 # Next
+
+# 6.5.0
 - Added `queue` parameter to `request` and `sendRequest`. This open up option to use other queue instead of main queue for response callback.
 
 # 6.4.0